Run compiled code on start
We had our `start` command running via `babel-node` which... technically
works but is slower (and maybe less secure).

Instead we want our start command, which may be invoked in production,
to build and run from the transpiled version of the code.

We'll address the dev command in a future commit.