Add hostkey setup instructions
This tells the user how to generate a hostkey and also updates our env
example to explain how to pass that hostkey to the service.

Note that the `ssh2` package provides a list of host key formats buried
in its documentation: https://www.npmjs.com/package/ssh2

It listed `ssh-ed25519` as most preferrable so I am using that in the
instructions.

Issue #3