Add always-accepting authentication
As described in the [SSH2
spec](https://datatracker.ietf.org/doc/html/rfc4252#section-5.2) the
client is able to pass `none` as the authentication type and we are
supposed to respond with a list of valid authentication methods for our
service.

We plan on requiring a bearer token to be passed as the password, and in
particular we aren't going to be using keys for authentication.

For now, however, we're going to just accept any password provided.

Issue #3