Start sftp sessions on request
If an sftp event is triggered via an ssh session we need to begin an
sftp stream.

This creates the stream by accepting the request and then stubs out the
handlers (similar to how we stubbed out the SSH handlers).

None of the sftp event handlers actually do anything yet, but by
stubbing them out we know what needs to be populated.

Issue #3