Add logger output for connection flow
By adding the logger debug output it's a lot easier to know what is
going on in the ssh connection.

I've also added some links to the relevant documentation for each
handler.  For now it's linking to the SSH2 documentation though we will
also want to add links to the SSH2 specification where appropriate.