Support archive loading
This adds initial support for loading archives from the Permanent API
and providing them to SFTP clients.  It does a few things temporarily
which will be removed:

1. It creates hard-coded root directories; this is for the sake of
   demonstration and will become dynamic.
2. It uses an environment variable for authentication; this will be
   replaced with the ability for clients to authenticate.
3. It uses very simple logic for directory / file detection.  This needs
   to be made more intelligent somehow.

This commit also updates the version `ssh2` since this is the first time
we're using typescript.

One thing to highlight is that sftp event handlers MUST be synchronous.
This is because event handlers in general must be synchronous.  There is
more information on the GitHub discussion referenced at the end of this
commit [1].

With this commit it should be possible for an sftp client to connect to
this service and view the list of archive directories associated with
the configured account.

[1] https://github.com/DefinitelyTyped/DefinitelyTyped/discussions/61265