Add token generation documentation
Authentication for this project is still a developer task -- which is to
say this does not accept user input for authentication and a given
instance can only be used to demonstrate against a single (configured)
user's account.

Until we leave that state we need to know how to generate an
authentication token.