Load folders from Permanent
We were previously spoofing any calls to access directories within an
archive folder by returning "empty" directory responses.  This was fine
for the demo but in the real world we need to actually interact with
Permanent to load the directory contents when a request is made.

In order to accomplish this we need to be able to map a path to a given
permanent folder ID. This requires us to know the Permanent ID for each
folder in the path.

This is achievable, but each lookup along the path requires an HTTP
request.  In order to try to minimize the number of lookups required we
want to be able to cache path => permanent object relationships so they
can be used for future paths.

The end result is that we want an instance to manage that memory, which
is why this commit introduces the concept of a `PermanentFileSystem`.
This class retains knowledge of any Permanent archive hierarchy that it
has been exposed to.

Issue #32 Use getFolder from the permanent SDK