Support record access
The SFTP service will now respond to requests to open, read, and close
files and will map those requests to appropriate calls to various
Permanent resources.

This implementation assumes that the file associated with a Permanent
"Record" is hosted by a service that is able to support Range headers in
HTTP requests [1].  It is also assumed that the `size` of the file as
returned by Permanent matches the size of the file on the CDN, as this
value is used to determine when to return an EOF signal to the SFTP
client.

[1] https://developer.mozilla.org/en-US/docs/Web/HTTP/Range_requests

Issue #35