Fix invalid filename bug
It turns out that when returning files (e.g. when getting the
contents of a directory) the value returned should be the file name NOT
the file path.  We were previously returning the path, which rclone was
tolerant of but I believe is actually against specification, and at the
very least is not supported by OpenSSH[1].

[1] https://www.ibm.com/docs/en/zos/2.2.0?topic=messages-fots2506