Update file type detection
The previous implementation of file detection was intentionally naive in
the spirit of rapid iteration.  It's time to actually detect whether
a given path is an item or a file based on the actual type according to
Permanent.

This will also resolve a related bug where folders on permanent with
a "." in the folder name would be incorrectly treated as files (and
therefore would not be downloaded).

Issue #54 'rclone copy' doesn't download all of my files