Use fileSystemCompatibleName for items
The latest version of the Permanent SDK exposes the
`fileSystemCompatibleName` for both Records and Folders, which is
intended to be used in filesystem contexts (instead of `name` or
`fileName`).

This will help the SFTP service prevent folder name collisions (fileName
was previously already serving the same purpose for Records).

Issue #60