Ensure unique file entries
The Permanent backend currently has a bug where it will sometimes return
duplicate records when getting folder contents.  That is a bug, but we
have the means to correct for it before the user is shown any kind of
warning by simply de-duplicating file entries within a folder before
returning them to the client.

Issue #56 Filter duplicate results from the server