Handle CHANNEL_EOF messages explicitly
There is an open PR in the upstream / ssh2 library related to properly
closing SFTP channels when a `CHANNEL_EOF` message is received[1].
Unfortunately that PR has not yet been merged.

Until that happens we need to handle the signal ourselves.
Unfortunately this requires us to access "private" attributes
(JavaScript doesn't support the concept of private attributes, so we are
able to do this...). This is, of course a terrible and horrible idea and
all of our tooling gets very upset about it.  As a result I had to
disable the tooling for the relevant line.

Once the PR is merged in upstream we should upgrade to it and delete the
contents of this commit.

Issue #45 rclone hangs after completion

[1] https://github.com/mscdex/ssh2/pull/1111