Migrate to 0.5.0 of permanent SDK
The new Permanent SDK has support for the creation of archive records.
It also renames `Record` to `ArchiveRecord` (Record had a naming
collision with the built in "Record" typescript utility type).  This
commit should not involve any functional changes, it just updates our
names to refer to "Archive Records" across the board.

Issue #77