Support file upload
This allows an sftp client to upload data to the Permanent virtual
filesystem.  Unfortunately the way sftp works there is no way of knowing
how large a file is going to be until after the handle is closed, and
unfortunately permanent requires us to know how large the file is before
it will allow us to generate a presigned post.

This means our solution is to write the file into temporary storage.
This is a potential vulnerability and should be seen as a short term
solution.

Other limitations of this implementation include:

1. Lack of support for editing files
2. Lack of support for updating modtime

All of these will need to be improved in future developments.

Issue #48 Implement file open / write handlers