Share filesystems per user
The previous implementation had one filesystem per connection.  This
created situations where a user's filesystem might update on one
connection but not update on another, which could result in
unpredictable behavior (e.g. if a folder was created in one connection
it might not be visible to another).

With this change those filesystems are now shared across all connections
for a given user.

Issue #76