Remove arrow functions for class methods
When an arrow function is used for a class method the function is not
a part of the prototype of the resulting object; rather, it is
a function that gets re-created every time an instance of the class is
built.  This is probably not going to result in major issues for us but
more generally it is an indication of why it is best practice to avoid
arrow functions for class methods.

In that spirit, this commit removes our instances of arrow function
class methods and adds the appropriate scope binding when those methods
are passed as handlers.

Issue #85 Class methods should be functions (not arrow functions)