Detect AWS status codes
It is possible that for various reasons AWS will fail to return file
content.  This could be if the URL is invalid, if the service is down,
etc.

Prior to this commit the SFTP service would not detect those messages
and would instead pass along the response body as though it was the file
content even if the response body was, for instance, error information.

Issue #88 If Permanent returns an invalid download URL our service
          returns invalid data to the client