Return lock status for processing records
When a record is created in Permanent it can take some time for the
Permanent backend to process the associated file.  During the processing
time that file is not available for read.  The SFTP protocol allows
a "lock" status to be returned for files that are currently locked by
another process, which seems apt for this situation.

The hope is that by returning a more specific status code to the
situation that SFTP clients are more likely to be able to appropriately
handle the file's status.

Issue #97 Return a more specific error code for archive records without
          populated files