Add cache bypasses
There are a few places where our caching logic was making it likely for
the cache to result in false misses.  The intent of the cache is pretty
specific: to avoid making API calls when crawling up the tree.

These changes really restrict the cache to those pieces, and should
prevent use of the cache for any item that is being directly accessed by
an SFTP call.  It also adds logic to refresh the cache one time whenever
there is a detected "miss" when retrieving a folder.

Issue #102