Wait for files to be processed
Just because a record exists in permanent doesn't mean that the record
has been processed. When accessing a file we need to make sure that the
file exists (e.g. has an "original file" and that original file has
a download URL and size).  This will wait for up to around 10 minutes
before giving up.

Issue #105