Fix parameter types for stat handlers
There were some mistakes in the parameter definition of various stat
handlers.  This wasn't manifesting in bugs yet, but it did mean that
data was incorrectly typed.  In the case of the FSTAT handler I believe
that handler would have failed since SFTP expects it to take a handler,
rather than a path.