Limit upload size to 5GB
AWS has a limit of 5GB uploads for single-part uploads, which is the
reason 5GB was selected here.

This does not necessarily prevent the risk of a DOS but it does require
a user to start multiple simultaneous uploads.

Issue #145 A user could create a DOS by uploading files that are too
large