Refactor handler logic
SFTP has the concept of a "handle", which is a pointer that clients will
send with requests to reference previously opened directories or files.
Our SftpSessionHandler had been storing various types of resources that
a handler could refer to in three different ways.  While this
technically "worked" the result was less organized than it should be.

In this commit we abstract the concept of a "Server Resource" and map
handles to various types of server resource.  Each SFTP event that works
with handles verify that the handle points to an appropriate type of
resource.

While making this change I also updated the way the server keeps track
of which file entries have already been "read" for a given handle.