Refactor temporary file management
We want to keep track of temporary files in order to delete "stale"
temporary files.  By incorporating a temporary file manager we (1)
encapsulate that management in a way that simplifies the SftpSessionHandler,
and (2) are able to track when temporary files have not been accessed in
a long period of time.

Issue #144 Temporary file storage can become "dead" if CLOSE is never called