Add `--inplace` flag instructions to README
In new versions of rclone it will try to upload to ".partial" files by
default and rename them to the proper destination filename after
uploading is fully completed. Unfortunately, we have not implemented the
rename command in SFTP yet, so this current upload flow doesn't work
with the sftp-service. Add instructions to the README to show users
which flag to pass into rclone to get it to work.