Use refresh token in 2FA flow
When refactoring authentication to use refresh tokens we missed the 2FA
flow. It's not clear that we should be supporting 2FA to begin with
since rclone doesn't support it, and ultimately the real solution is to
use keys instead of passwords for sftp authentication.

That being said, while it's here we should make sure it isn't broken!

Issue #298 2FA auth flow does not utilize refresh tokens