Update caches on creation
We were not updating our local caches when items were created; this was
not resulting in functional issues because of our cache-busting logic,
but it was resulting in non-optimal caching and an increase in API calls
as a result of being unable to rely on the cache.

Specifically, rclone verifies files immediately after creation by
checking their stats.  This meant that every time a file was uploaded it
would result in a preventable API call to `/folder/getWithChildren`.
There were some circumstances where a similar unnecessary call could
happen after folder creation.

Issue #306 Update cache after file / folder creation