Properly update archiveFolder cache
Archive folders are currently stored in a separate cache from other
folders. This may not actually be the correct design decision, but since
it is the current implementation we do at least want to make sure that
the cache optimization related to updating caches with information about
newly created items applies to those folders as well.

Issue #306 Update cache after file / folder creation