Bypass cache busting when opening files for write
Our cache buster logic will generally bypass cached values in the event
of a negative result -- that is, if a file or folder is being opened but
doesn't exist in the cache, it will attempt to load that file or folder
directly from the API instead of assuming the cache is accurately
reflecting it's lack of existence.

This is a reasonable tactic, but in the event of file upload it
predictably results in a lot of API calls (since, generally speaking, if
a file is being uploaded that's because it doesn't exist!).

Those calls add up, and so this logic changes the "check" when opening
a file for write purposes.  The new check doesn't actually attempt to
load the file, but instead directly looks at whether the file exists in
the current cache (without busting that cache in the event it doesn't).

Note that in the event that the file's *parent directory* doesn't exist
in the cache, we do need to try to load that parent directory.  This is
for two reasons: (1) we need the parent directory to exist and (2) if
the parent directory hasn't loaded at all this dramatically affects our
trust in the (non-existent) cache of that parent directory.

Issue #308 Consider bypassing cache buster when opening a file for write