Update switch statements to obey linting rules
An update to eslint includes a rule that the values in the case
statements of a switch must match the type of the value in the
predicate. We have some switch statements that disobey this, because the
cases use enum values, while the predicates take a value from a
fusionAuth response with type number. We can't change the types of the
fusionAuth response (because we want to use their provided client), so
this commit fixes the problem by calling .valueOf() on the enum values
in the case statements, the result of which will have type number.