Remove wait for FileVOs to be populated
Now that FileVOs are populated in calls to `registerRecord`, we no
longer have to wait for files to process for an arbitrary amount of time
before being able to see the FileVOs field. Remove the code that waits
for this field to be populated since now it's theoretically creating a 1
second per file delay in the upload process.

PER-9417: Remove "wait for processing" step in sftp-service