Move Sentry initialization to instrument.js
In their v8 release, Sentry added automatic instrumentation for a
variety of node packages. For this to work, Sentry must be initialized
before any other import. To achieve this, this commit moves Sentry
initialization into an instrument.js file, which is now the first import in
index.js, as recommended in the Sentry release notes.