Opt to fail on duplicate names when creating records and folders
By default, Permanent allows the creation of folders and records with
duplicate names, creating disambiguated "download names" under the hood.
However, an SFTP server should not be able to create items with
duplicate names, nor should it create items with different names than
what it was asked to create. To this end, we added a failOnDuplicateName
flag to the Permanent endpoints that create records and folders. This
commit updates the SFTP service to call those endpoints with that flag
set to true.