Use `getItemType` to validate path type
When handling an `OPEN` event for opening a file we were verifying that
the path was a file (rather than a directory) in a very strange way.
I say this having been the person who wrote it...

We tried opening it as a directory, and then expected it to fail.  This
is expensive and indirect compared to simply loading the item type using
the PermanentFileSystem's `getItemType` method.

Switching to `getItemType` appears to significantly speed up the
service.  This makes sense as cache-busting calls to the API are
triggered when attempting to load a folder path that doesn't exist (and,
in this case the folder SHOULDN'T exist).

Issue #577 Use getItemType to check if a path is a directory when handling file open