Only return NO_SUCH_FILE if the file wasn't found
The previous logic would always return a `NO_SUCH_FILE` status code if
the PermanentFileSystem threw an error when attempting to load a file or
folder.

In reality there are a variety of reasons that PermanentfileSystem might
throw an error -- one is that the resource doesn't exist but another
example we saw was if API calls failed for other reasons (e.g. 504
Gateway Timeout, or 500 errors / bugs on the backend).

Returning a more accurate status code won't necessarily result in
a resolved user experience, but it should at least allow SFTP clients to
behave in a more appropriate way (e.g. potentially retrying the call)

Issue #590 Ensure FAILURE if api calls fail