Remove duplicate error type
We had `ResourceDoesNotExistError` and `NotFoundError`

Both of these appear to mean the same thing, and without clear
differentiation it is just confusing to have two of them.  This was
leading to issues where our SFTP session handler had logic to look for
one of the errors to return a NotFound status, but didn't check for
both.

This was leading to an issue where folder creation via sftp would fail.

Issue #607 Errors when attempting to create a folder