Normalize path and filesystem language
We want to be as clear and precise as possible about when we are
referring to objects in the Permanent domain (i.e. archives, archive
records, and folders) vs the virtual file system domain (files and
directories).

This attempts to be more explicit that paths are referring to the
virtual file system and we are loading filesystem object types
/ attributes, rather than permanent object types / attributes.