Normalize errors related to permission
We were using a single error for a range of error types that included
errors related to what a user is allowed to do as well as function input
validation checks.

This commit creates a single (better named) error to signal when an
error is thrown to indicate a lack of user permission to perform
a function.