Properly handle `SSH_FXP_CLOSE` on directory handles
Directories need to be closed too!  We were returning a failure status
but this was incorrect.  I don't believe this was manifesting in any
bugs, but it was definitely resulting in `FAILURE` entires in log files
that did not belong.

Issue #619 SSH_FXP_CLOSE should support closing directory handles too