Add a universal path check
The PermanentFileSystem has some constraints which can be checked at the
path level.  For example, only a single root directory exists
(`/archives`) and therefore any attempt to read or write to a path
outside of that directory structure needs to be invalidated.

We have a series of path categorization functions (e.g.
`isArchivePath`), but in some cases we were not properly checking to
ensure that the path existed in `/archives/*`

This adds a new function which will validate if a path is one of the
valid path categories, which will make it easier for us to defensively
halt processing of any paths that might be incorrectly constructed (e.g.
`/foo/blah/bar.txt`)

Issue #615 Ensure non/archives paths are properly rejected