Return `NO_SUCH_FILE` for `SSH_FXP_OPENDIR`
When attempting to open a directory that does not exist, our SFTP
service was returning a `FAILURE` instead of the more correct
`NO_SUCH_FILE`.

The SFTP specification generally says that FAILURE should only be used
when a more precise error does not exist.

The result could be that clients would not know that a directory needed
to be created, which could exacerbate certain error states in a client
/ now allow appropriate recovery.

Issue #625 Return NO_SUCH_FILE for the SSH_FXP_OPENDIR operation