Respond with `NO_SUCH_FILE` when opening a file for write
If a directory path does not exist we should reply with `NO_SUCH_FILE`
instead of a generic `FAILURE` when a file is being opened for write.

This allows the client to recover by attempting to create the parent
directory tree.

Issue #627 Respond with NO_SUCH_FILE when attempting to open a file for write in a nonexistent directory