Return specific status codes for `SSH_FXP_MKDIR`
We were returning a generic `FAILURE` for any type of issue when
a client tried to make a directory.  We need to be more specific
and indicate when the parent directory doesn't exist vs when
the user is not allowed to create the directory due to a permissions
issue.

Issue #628 Respond with more specific failure codes on SSH_FXP_MKDIR