Consistently return `PERMISSION_DENIED` correctly
Every SFTP operation that is interacting with the PermanentFileSystem
has the potential to receive a PermissionDeinedError (e.g. if attempting
to write to an invalid path, or attempting to interact with a resource
that they do not have the right to interact with).

In most cases this was resulting in a generic `FAILURE` status response
to the SFTP client, but we can do better.

There were also some instances where a `PERMISSION_DENIED` status was
being returned when a `FAILURE` actually is more appropriate (e.g.
attempting to overwrite an existing file is less a permission error so
much as a general failure state).

Issue #631 Return PERMISSION_DENIED when appropriate on all SFTP operations