Retry on 429 errors
We have found that FusionAuth can rate limit the Stela API if there are
too many requests in rapid succession for a given token.  Stela is now
exposing the 429 error when that happens and we should use that as an
opportunity to retry with exponential decay.

This commit just adds 429 to the list of valid retry codes for now.
There is an upcoming SDK change which will also allow us to configure
more retries that happen over a longer period of time.

Issue #650 Support retry on 429 errors