Retry for longer
The new SDK supports more configuration for retry logic; we have found
that unrecoverable failures results in cascading faliure states within
rclone and so it is worth retrying for much longer than the SDK default
in the hopes that things resolve.

This is particularly important if we're getting rate limit errors.

Issue #650 Support retry on 429 errors