Enable `@typescript-eslint/class-methods-use-this`
Enabling this lint rule exposed the fact that we were using a global
logger instead of an injectable logger in these classes.  This is true
elsewhere as well, but fixing it here is still an improvement and also
allows us to adhere to this rule.

Issue #675 Re-enable disabled lint rules