Add initial unit tests
This narrowly scoped commit adds our first unit tests.  This particular
function was picked just because it was a good candidate and the real
contribution here is moving from 0% coverage to >0% coverage.

This allows us to ensure our tests actually run, locally and in CI, and
sets the stage for additional unit tests in the future.

Issue #33 We need tests